Update Code 907. Sum of Subarray Minimums.cpp
typo

+++ LeetCode/Medium/Stack/Code 907. Sum of Subarray Minimums.cpp
@@ -25,6 +25,6 @@ public:
             result %= MOD;
         }
 
-        return static_cast<int>(result);
+        return result;
     }
 };
